+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/MavenIntegrationHarness.java
@@ -136,8 +136,12 @@ public class MavenIntegrationHarness extends ResourceHarness {
 		writePom(groupWithSteps("cpp", steps));
 	}
 
-	protected void writePomWithTypescriptSteps(String... steps) throws IOException {
-		writePom(groupWithSteps("typescript", including("**/*.ts"), steps));
+	protected void writePomWithJavascriptSteps(String includes, String... steps) throws IOException {
+		writePom(groupWithSteps("javascript", including(includes), steps));
+	}
+
+	protected void writePomWithTypescriptSteps(String includes, String... steps) throws IOException {
+		writePom(groupWithSteps("typescript", including(includes), steps));
 	}
 
 	protected void writePomWithSqlSteps(String... steps) throws IOException {